TST-87: Added Party Details to Partybar
The partybar would not update with the hosts information
even after a selection was made.

The party bar now displays information about the selected party.
It has been made into an accordion dropdown with options for the
user to view their current party details or make a different
selection.

Changes were made to both the SCSS and HTML. The SCSS was not very
efficiently strucutred and has been refactored.