TST-87: Fixed Conflicts with Partybar Move
The partybar was moved out of the header, but none of the styling
information or scripts were changed to reflect that.

The partybar has been changed to an html component and the styling
has been moved to the 'find-a-party.scss' file. This ensures that
the styles will be present no matter where the element is on the
page.